backend/requirements.txt
django==4.2.9
djangorestframework==3.14.0
pymongo[srv]==4.6.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
